+++ src/main/java/LengBD/domain/AsignacionListadoDTO.java
@@ -11,7 +11,7 @@ public class AsignacionListadoDTO {
     private LocalDate fechaInicio;
     private LocalDate fechaFinal;
     private String motivo;
-    private Long cedula;             
+    private Integer cedula;             
     private Integer idInstrumento;    
     private String nombreEstado;
     private Integer idEstado;
